RemoteControl: never have more than one pending refresh event. Mostly fixes #2518.
Feeding the server with lots of commands (as described in #2518) no longer
overloads the application's event loop with user events.
However, there are still noticeable slowdowns on the display, most
probably caused by the fact that the network reception code runs in the main
event loop. Running the server in its own thread may be a solution.