RemoteControl: run each Connection in its own thread. Refs #2518.
This is a significant rework of the module. Although existing classes
keep the same role, there is now one thread per ClientConnection so that
the network events generated by incoming commands are not dispatched to
the main event loop. This minimizes the impact on the refresh rate of the
display.
This commit also makes -tremotecontrol less verbose, by not logging
incoming commands anymore. Instead, two new trace names have beed added:
 * -tremotecontrolcmd logs all incoming commands,
 * -tremotecontrolcode logs when XL code is updated.